pyproject.toml
[build-system]
requires = ["maturin>=1.5,<2"]
build-backend = "maturin"

[project]
name = "goop-veil"
version = "0.1.0"
description = "WiFi privacy defense system — detection, passive shielding, and active countermeasures"
license = {text = "Apache-2.0"}
requires-python = ">=3.11"
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Rust",
    "Topic :: Security",
    "Topic :: System :: Networking",
]
dependencies = ["pydantic>=2.0,<3"]

[project.optional-dependencies]
cli = ["typer>=0.9,<1", "rich>=13,<14"]
mcp = ["mcp>=1.2,<2"]
active = ["pyserial>=3.5,<4"]
dashboard = ["fastapi>=0.109", "uvicorn>=0.27"]
router = ["paramiko>=3.0,<4"]
router-api = ["httpx>=0.25"]
traffic = ["aiohttp>=3.9"]
legal = ["jinja2>=3.1"]
smart-home = ["zeroconf>=0.80"]
mitigation = ["goop-veil[router,router-api,traffic,legal]"]
dev = ["pytest>=7", "pytest-asyncio>=0.23", "ruff>=0.3", "mypy>=1.8"]
all = ["goop-veil[cli,mcp,active,dashboard,mitigation,smart-home,dev]"]

[project.urls]
Homepage = "https://github.com/kobepaw/goop-veil"
Repository = "https://github.com/kobepaw/goop-veil"

[project.scripts]
goop-veil = "goop_veil.cli:main"

[tool.maturin]
python-source = "python"
module-name = "goop_veil._core"
features = ["extension-module"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
